please grade homework4 osx

+++ src/gamedata/Actor.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <stdint.h>
 
 #include "glm/glm.hpp"
 
@@ -23,7 +24,7 @@ public:
 	std::string nearbyDialogue;
 	std::string contactDialogue;
 
-	uint64_t id = 0;
+	int id = 0;
 	bool triggeredScoreUp = false;
 
 	Actor(ActorProps& props)
